+++ webapp/src/app/shared/widgets/widget.module.ts
@@ -29,6 +29,8 @@ import {AddEventService} from './event-widget/event-editor/add/add-event.service
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { EnemyTypeWidgetComponent } from './enemy-type-widget/enemy-type-widget.component';
 import { EnemyTypeWidgetOverlayComponent } from './enemy-type-widget/enemy-type-overlay/enemy-type-overlay.component';
+import {LangLabelWidgetComponent} from './langlabel-widget/langlabel-widget.component';
+import {AngularResizeEventModule} from 'angular-resize-event';
 
 const COMPONENTS = [
 	StringWidgetComponent,
@@ -44,6 +46,7 @@ const COMPONENTS = [
 	CharacterWidgetComponent,
 	PersonExpressionWidgetComponent,
 	EnemyTypeWidgetComponent,
+	LangLabelWidgetComponent
 ];
 
 const PRIVATE_COMPONENTS = [
@@ -63,7 +66,8 @@ const PRIVATE_COMPONENTS = [
 		AngularDraggableModule,
 		DragDropModule,
 		ScrollingModule,
-		SharedModule
+		SharedModule,
+		AngularResizeEventModule,
 	],
 	providers: [
 		EventRegistryService,