+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- New editor for `langlabel` properties.
+
 ### Fixed
 - Typo in settings panel.
 - Fixed compatibility with the newer Electron versions.